+++ Transfers/uploadfiletransferapi.h
@@ -48,8 +48,8 @@ public:
 	~UploadFileTransferAPI();
 	virtual void start();
 	virtual void cancel();
-	virtual int getTransferedBytes();
-	virtual int getTotalBytes();
+	virtual long long getTransferedBytes();
+	virtual long long getTotalBytes();
 };
 
 #endif //H_UPLOADFILETRANSFERAPI
